+++ src/lib.rs
@@ -7,8 +7,8 @@ pub mod vars;
 #[deprecated]
 pub use vars::EnvEvalable;
 pub use vars::{
-    CwdGuard, EnvDict, EnvEvaluable, Mutability, OriginDict, OriginValue, UpperKey,
+    CwdGuard, EnvChecker, EnvDict, EnvEvaluable, Mutability, OriginDict, OriginValue, UpperKey,
     ValueConstraint, ValueDict, ValueObj, ValueType, ValueVec, VarCollection, VarDefinition,
-    VarToValue, find_project_define, find_project_define_base, find_project_root,
-    find_project_root_from, setup_start_env_vars,
+    VarToValue, extract_env_var_names, find_project_define, find_project_define_base,
+    find_project_root, find_project_root_from, setup_start_env_vars,
 };